Move AnySpi and DataMode
