+++ esp-hal/CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `AnyI2c` has been moved from `esp_hal::i2c` to `esp_hal::i2c::master` (#3627)
+- `AnySpi` has been moved from `esp_hal::spi` to `esp_hal::spi::master` and `esp_hal::spi::slave` (#3627)
+- `DataMode` has been moved from `esp_hal::spi` to `esp_hal::spi::master` (#3627)
 
 ### Fixed
