+++ esp-metadata/devices/esp32c2.toml
@@ -80,6 +80,7 @@ bus_timeout_is_exponential = true
 
 [device.spi_master]
 status = "supported"
+instances = [{ name = "spi2" }]
 
 [device.timergroup]
 instances = [{ name = "timg0" }]
