+++ esp-metadata/devices/esp32c6.toml
@@ -131,6 +131,7 @@ channel_ram_size = 48
 
 [device.spi_master]
 status = "supported"
+instances = [{ name = "spi2" }]
 
 [device.timergroup]
 instances = [{ name = "timg0" }, { name = "timg1" }]
