+++ esp-metadata/devices/esp32s3.toml
@@ -126,6 +126,7 @@ channel_ram_size = 48
 
 [device.spi_master]
 status = "supported"
+instances = [{ name = "spi2" }, { name = "spi3" }]
 
 [device.timergroup]
 timg_has_timer1 = true
