+++ hil-test/tests/spi_half_duplex_write.rs
@@ -13,9 +13,8 @@ use esp_hal::{
     gpio::interconnect::InputSignal,
     pcnt::{Pcnt, channel::EdgeMode, unit::Unit},
     spi::{
-        DataMode,
         Mode,
-        master::{Address, Command, Config, Spi, SpiDma},
+        master::{Address, Command, Config, DataMode, Spi, SpiDma},
     },
     time::Rate,
 };
