+++ hil-test/tests/spi_half_duplex_write_psram.rs
@@ -17,9 +17,8 @@ use esp_hal::{
     gpio::interconnect::InputSignal,
     pcnt::{Pcnt, channel::EdgeMode, unit::Unit},
     spi::{
-        DataMode,
         Mode,
-        master::{Address, Command, Config, Spi, SpiDma},
+        master::{Address, Command, Config, DataMode, Spi, SpiDma},
     },
     time::Rate,
 };
